fix(ci): clear 32 ruff lint errors blocking CI green
Auto-fixed 25 (import sorting, unused imports, asyncio.TimeoutError →
TimeoutError, RUF022 __all__ sort) via ruff check --fix; manually
addressed the remaining 7:

- src/hal0/api/routes/slots.py: replace ambiguous U+2212 with hyphen
  in the docstring (RUF002)
- src/hal0/auth/tokens.py: try/except/pass → contextlib.suppress
  on the metadata-write best-effort path (SIM105)
- src/hal0/hardware/probe.py: drop redundant int() around round()
  (RUF046)
- src/hal0/providers/llama_server.py: hoist _gpu import to the top
  block, restoring E402 compliance
- tests/api/test_slots_routes.py: try/except/pass → contextlib.suppress
  on cancelled-task cleanup (SIM105)
- tests/auth/test_tokens.py: rename unused `tok` to `_tok` (RUF059)
- tests/registry/test_curated.py: tighten blind pytest.raises(Exception)
  to pydantic.ValidationError (B017)

No logic changes — `ruff check src tests` is clean. Pre-existing
unrelated test failure in tests/providers/test_comfyui.py
(image_ref returns a sha256 digest but the test expects ":v1") is
left untouched; tracked separately.

+++ src/hal0/hardware/probe.py
@@ -28,9 +28,9 @@ from pathlib import Path
 
 import structlog
 
-from hal0.errors import Hal0Error
 from hal0.config import paths as _paths
 from hal0.config.schema import GPUInfo, HardwareInfo, NPUInfo
+from hal0.errors import Hal0Error
 
 log = structlog.get_logger(__name__)
 
@@ -197,7 +197,7 @@ def _dmidecode_host_ram_mb() -> int | None:
                     total_mb += n * mult
     if total_mb <= 0:
         return None
-    return int(round(total_mb))
+    return round(total_mb)
 
 
 def _derive_unified_memory_mb(ram_mb: int, gpu: GPUInfo | None) -> int:
